3 Streamlined Participant Addition
Streamlined the participant addition by creating an inline form for it where all details for the participant are defined before adding it.  Eliminated participant locking so that the user doesn't have to click down to the participant after adding it to enter details.

Removed all calls to console.log because they shouldn't be in the code base.